chore: moved generators
