+++ packages/cli/src/commands/generate/generators/index.ts
