+++ packages/cli/src/commands/generate/generators/prisma.ts
